automation/python/requirements.txt
netmiko>=4.2.0
napalm>=4.1.0
paramiko>=3.3.0
jinja2>=3.1.0
pyyaml>=6.0
rich>=13.0
click>=8.1.0
deepdiff>=6.0
tabulate>=0.9.0
